+++ change_log.md
@@ -63,4 +63,5 @@
     - Fixed bug: that when *actionNeedToBeEnded* is enabled that the Boolean whether the participant is moving or not was not working.
 - Version 3.0.0
     - Fixed introduced bug that seession was not added to countdown. 
-    - Fixed further bugs that were introduced by changing the items: 1. fixed dimensions of the UXF UI. 2. fixed the animation of the trial starter. 
+    - Fixed further bugs that were introduced by changing the items: 1. fixed dimensions of the UXF UI, 2. fixed the animation of the trial starter (removing unsupported VFX), 3. aligned the objects in the welcome screen, 4. fixed the experiment controller script and 5. remove the word icosahedron. 
+    - Add study dictionary method, change UXF UI for WebGL and change UXF UI controll script