+++ change_log.md
@@ -70,3 +70,4 @@
         - Fixed the input field font size.
     - Fixed FPS issue that average is not caluclated correctly in WebGL.
     - Removed reflection probes, scene lighting data. Also move settings filers to tracked folder.
+    - Baked lights even if benefit is not entirely clear.