+++ change_log.md
@@ -72,3 +72,4 @@
     - Removed reflection probes, scene lighting data. Also move settings filers to tracked folder.
     - Baked lights even if benefit is not entirely clear. 
     - Bakad again but didn't remove the reflection probes as it seems important for the objects rendering correctly. 
+    - Clean up stuff from old version like prefabs and samples from UXF.