+++ change_log.md
@@ -71,7 +71,8 @@
     - Fixed FPS issue that average is not caluclated correctly in WebGL.
     - Removed reflection probes, scene lighting data. Also move settings filers to tracked folder.
     - Baked lights even if benefit is not entirely clear. 
-    - Bakad again but didn't remove the reflection probes as it seems important for the objects rendering correctly. 
+    - Baked again but didn't remove the reflection probes as it seems important for the objects rendering correctly. 
     - Clean up stuff from old version like prefabs and samples from UXF.
     - Number of small changes: 1. Renamed destroyAllCubesAtTrialEnd function, 2. Update file logo, 3. add object index to log. 
     - Create new basic layout with minimum distance betwee objects.
+    - Added mode 'standard' & 'fMRI'.